+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.0.12
 Release TBD
   * Move Message and JobQueue functions into classes.
+  * Enable SmartResizer to return full image urls and skip lightbox tags.
+  * Lower PHP requirement to 7.2.0.
 
 ## 1.0.11
 Release 2020-07-27